refactor: feature-rooms depends on core-rooms, list is reactive
