+++ shared/feature-rooms/impl/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/rooms/impl/di/FeatureRoomsImplModule.kt
@@ -1,13 +1,7 @@
 package dev.nonoxy.residetrack.feature.rooms.impl.di
 
 import dev.nonoxy.residetrack.common.coroutines.CoroutineDispatchers
-import dev.nonoxy.residetrack.feature.rooms.api.repository.RoomsRepository
 import dev.nonoxy.residetrack.feature.rooms.api.store.RoomsStore
-import dev.nonoxy.residetrack.feature.rooms.impl.data.RoomsRepositoryImpl
-import dev.nonoxy.residetrack.feature.rooms.impl.data.mappers.RoomMapper
-import dev.nonoxy.residetrack.feature.rooms.impl.data.mappers.RoomMapperImpl
-import dev.nonoxy.residetrack.feature.rooms.impl.data.mappers.StudentMapper
-import dev.nonoxy.residetrack.feature.rooms.impl.data.mappers.StudentMapperImpl
 import dev.nonoxy.residetrack.feature.rooms.impl.domain.RoomsStoreFactory
 import dev.nonoxy.residetrack.feature.rooms.impl.domain.seed.RoomsSeedInitializer
 import dev.nonoxy.residetrack.core.initializer.Initializer
@@ -18,22 +12,6 @@ import org.koin.dsl.module
 val featureRoomsImplModule = module {
 
     factoryOf(::RoomsSeedInitializer) bind Initializer::class
-
-    factory<StudentMapper> {
-        StudentMapperImpl()
-    }
-
-    factory<RoomMapper> {
-        RoomMapperImpl(studentMapper = get())
-    }
-
-    factory<RoomsRepository> {
-        RoomsRepositoryImpl(
-            roomDao = get(),
-            roomMapper = get(),
-            studentMapper = get(),
-        )
-    }
 
     factory<RoomsStore> {
         RoomsStoreFactory(
